requirements.txt
PyPDF2>=3.0.0  # PDF text extraction
python-docx>=1.1.0  # DOCX file handling
openai>=1.0.0  # LLM client base

python-dotenv>=1.0.0  # For environment variables
pytest>=7.0.0  # Testing framework
mypy>=1.0.0  # Static type checking

jsonschema>=4.0.0  # For response validation
typing-extensions>=4.0.0  # For TypedDict support
python-dotenv>=0.19.0
tqdm>=4.0.0  # Optional progress bars
colorama>=0.4.0  # Colored console output